JAMES-2586 - CLEAN CODE -  Guice binding for Postgres User Repository modules
- Making it easier to install the binding when a user chooses this option.
