+++ server/container/guice/postgres-common/src/main/java/org/apache/james/modules/data/PostgresDelegationStoreModule.java
@@ -19,22 +19,12 @@
 
 package org.apache.james.modules.data;
 
-import org.apache.commons.configuration2.ex.ConfigurationException;
-import org.apache.james.backends.postgres.PostgresModule;
-import org.apache.james.server.core.configuration.ConfigurationProvider;
 import org.apache.james.user.api.DelegationStore;
 import org.apache.james.user.api.DelegationUsernameChangeTaskStep;
 import org.apache.james.user.api.UsernameChangeTaskStep;
-import org.apache.james.user.lib.UsersDAO;
 import org.apache.james.user.postgres.PostgresDelegationStore;
-import org.apache.james.user.postgres.PostgresUserModule;
-import org.apache.james.user.postgres.PostgresUsersDAO;
-import org.apache.james.user.postgres.PostgresUsersRepositoryConfiguration;
 
 import com.google.inject.AbstractModule;
-import com.google.inject.Provides;
-import com.google.inject.Scopes;
-import com.google.inject.Singleton;
 import com.google.inject.multibindings.Multibinder;
 
 public class PostgresDelegationStoreModule extends AbstractModule {
@@ -45,18 +35,5 @@ public class PostgresDelegationStoreModule extends AbstractModule {
 
         Multibinder.newSetBinder(binder(), UsernameChangeTaskStep.class)
             .addBinding().to(DelegationUsernameChangeTaskStep.class);
-
-        bind(PostgresUsersDAO.class).in(Scopes.SINGLETON);
-        bind(UsersDAO.class).to(PostgresUsersDAO.class);
-
-        Multibinder<PostgresModule> postgresDataDefinitions = Multibinder.newSetBinder(binder(), PostgresModule.class);
-        postgresDataDefinitions.addBinding().toInstance(PostgresUserModule.MODULE);
-    }
-
-    @Provides
-    @Singleton
-    public PostgresUsersRepositoryConfiguration provideConfiguration(ConfigurationProvider configurationProvider) throws ConfigurationException {
-        return PostgresUsersRepositoryConfiguration.from(
-            configurationProvider.getConfiguration("usersrepository"));
     }
 }
