+++ server/container/guice/postgres-common/src/main/java/org/apache/james/modules/data/PostgresUsersRepositoryModule.java
@@ -19,21 +19,46 @@
 
 package org.apache.james.modules.data;
 
+import org.apache.commons.configuration2.ex.ConfigurationException;
+import org.apache.james.backends.postgres.PostgresModule;
 import org.apache.james.server.core.configuration.ConfigurationProvider;
 import org.apache.james.user.api.UsersRepository;
+import org.apache.james.user.lib.UsersDAO;
+import org.apache.james.user.postgres.PostgresUserModule;
+import org.apache.james.user.postgres.PostgresUsersDAO;
 import org.apache.james.user.postgres.PostgresUsersRepository;
+import org.apache.james.user.postgres.PostgresUsersRepositoryConfiguration;
 import org.apache.james.utils.InitializationOperation;
 import org.apache.james.utils.InitilizationOperationBuilder;
 
 import com.google.inject.AbstractModule;
+import com.google.inject.Provides;
 import com.google.inject.Scopes;
+import com.google.inject.Singleton;
+import com.google.inject.multibindings.Multibinder;
 import com.google.inject.multibindings.ProvidesIntoSet;
 
 public class PostgresUsersRepositoryModule extends AbstractModule {
+
+    public static AbstractModule USER_CONFIGURATION_MODULE = new AbstractModule() {
+        @Provides
+        @Singleton
+        public PostgresUsersRepositoryConfiguration provideConfiguration(ConfigurationProvider configurationProvider) throws ConfigurationException {
+            return PostgresUsersRepositoryConfiguration.from(
+                configurationProvider.getConfiguration("usersrepository"));
+        }
+    };
+
     @Override
     public void configure() {
         bind(PostgresUsersRepository.class).in(Scopes.SINGLETON);
         bind(UsersRepository.class).to(PostgresUsersRepository.class);
+
+        bind(PostgresUsersDAO.class).in(Scopes.SINGLETON);
+        bind(UsersDAO.class).to(PostgresUsersDAO.class);
+
+        Multibinder<PostgresModule> postgresDataDefinitions = Multibinder.newSetBinder(binder(), PostgresModule.class);
+        postgresDataDefinitions.addBinding().toInstance(PostgresUserModule.MODULE);
     }
 
     @ProvidesIntoSet
@@ -42,4 +67,5 @@ public class PostgresUsersRepositoryModule extends AbstractModule {
             .forClass(PostgresUsersRepository.class)
             .init(() -> usersRepository.configure(configurationProvider.getConfiguration("usersrepository")));
     }
+
 }
