+++ server/data/data-postgres/src/main/java/org/apache/james/user/postgres/PostgresDelegationStore.java
@@ -47,7 +47,7 @@ public class PostgresDelegationStore implements DelegationStore {
         }
     }
 
-    private PostgresUsersDAO postgresUsersDAO;
+    private final PostgresUsersDAO postgresUsersDAO;
     private final UserExistencePredicate userExistencePredicate;
 
     @Inject
